requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.12.4
graphene==2.1.8
graphene-django==2.12.0
graphql-core==2.3.2
promise==2.3
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
